refactoring to make uniform main program
