+++ Abstract Factory pattern - 2/DatabaseSqlServer.cs
@@ -9,12 +9,12 @@ namespace BridgeTechWhizz
     {
         public DatabaseSqlServer()
         {
-            CreateConnectionAndCommand(); // calling factory method - Second factory method give the connection and command
+            CreateConnection(); // calling factory method - Second factory method give the connection and command
         }
         public DbConnection Connection { get; set; }
 
 
-        public void CreateConnectionAndCommand() // factory method to decide connection and command type
+        public void CreateConnection() // factory method to decide connection and command type
         {
             Connection = new SqlConnection();
         }
