+++ Abstract Factory pattern - 2/IDatabase.cs
@@ -6,9 +6,8 @@ namespace BridgeTechWhizz
     public interface IDatabase
     {
         DbConnection Connection { get; set; }
-        //DbCommand Command { get; set; }
 
-        void CreateConnectionAndCommand();
+        void CreateConnection();
 
         void PrintConnection();
     }
