+++ Abstract Factory pattern/0002 - Creational - Abstract Factory Pattern - 1.csproj
@@ -53,11 +53,14 @@
     <Compile Include="Deer.cs" />
     <Compile Include="Hunter.cs" />
     <Compile Include="Lion.cs" />
-    <Compile Include="Main.cs" />
+    <Compile Include="_Main.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Rabbit.cs" />
     <Compile Include="Victim.cs" />
     <Compile Include="Wolf.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="Readme.txt" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
