+++ Abstract Factory pattern/Readme.txt
@@ -1,0 +1,9 @@
+﻿Abstract factory pattern
+-------------------------
+
+Factory inside factory
+
+Animal world example
+------------------------------
+Here we are trying print food chain (Who is the Hunter and who is the Victim) of two differnt continent
+
