+++ Adapter Desig Pattern/Realworld/0005 - Structural - Adapter - PAN Adhaar Import.csproj
@@ -54,7 +54,7 @@
     <Compile Include="PersonAdhaar.cs" />
     <Compile Include="PersonLocal.cs" />
     <Compile Include="PersonPAN.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="_Main.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
