+++ Adapter Pattern Robort Example/0006 - Structural - Adapter pattern robot example.csproj
@@ -45,7 +45,7 @@
   <ItemGroup>
     <Compile Include="EnemyRobot.cs" />
     <Compile Include="EnemyRobotAttackerAdapter.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="_Main.cs" />
     <Compile Include="EnemyTank.cs" />
     <Compile Include="IEnemyAtacker.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
