+++ Chain Of Resp Design Pattern/0010 - Behavioral - ChainOfResponsibility.csproj
@@ -49,7 +49,7 @@
     <Compile Include="Approver.cs" />
     <Compile Include="Director.cs" />
     <Compile Include="President.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="_Main.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Purchase.cs" />
     <Compile Include="VicePresident.cs" />
