+++ Facade Design Pattern/0007 - Structural - Facade Design Pattern.csproj
@@ -53,7 +53,7 @@
     <Compile Include="Order.cs" />
     <Compile Include="PriceList.cs" />
     <Compile Include="ProductDetails.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="_Main.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PurchaseItem.cs" />
     <Compile Include="ShoppingCart.cs" />
