+++ Factory Method Design pattern/Realworld/_Main.cs
@@ -1,0 +1,46 @@
+﻿// Client class
+
+using System;
+
+namespace BridgeTechWhizz
+{
+    public class MainApp
+    {
+        static void Main()
+        {
+            Console.WriteLine("Create Resume for Higher Education (Creation of *Higher Education Resume* is happening in a separate factory class)\n------------------------");
+            Document_Abstract resumeEducation2 = new DocumentFactory().GetDocument(1);
+            foreach (var resumeEducationPage in resumeEducation2.Pages)
+            {
+                Console.WriteLine(resumeEducationPage.PageContent);
+            }
+
+            Console.WriteLine("\n\n");
+            Console.WriteLine("Create Resume for Job (Creation of *Job Resume* is happening in a separate factory class)\n------------------------");
+            Document_Abstract resumeJob2 = new DocumentFactory().GetDocument(2);
+            foreach (var resumeJobPage in resumeJob2.Pages)
+            {
+                Console.WriteLine(resumeJobPage.PageContent);
+            }
+            
+            Console.WriteLine("\n\n\n\n");
+
+            Console.WriteLine("Create Resume for Higher Education (Creation of *Document pages of Higher Education Resume* is happening inside the concrete object)\n------------------------");
+            Document_Abstract resumeEducation1 = new Document_HigherEducationResume();
+            foreach (var resumeEducationPage in resumeEducation1.Pages)
+            {
+                Console.WriteLine(resumeEducationPage.PageContent);
+            }
+
+            Console.WriteLine("\n\n");
+            Console.WriteLine("Create Resume for Job (Creation of *Document pages of Job Resume* is happening inside the concrete object)\n------------------------");
+            Document_Abstract resumeJob1 = new Document_JobResume();
+            foreach (var resumeJobPage in resumeJob1.Pages)
+            {
+                Console.WriteLine(resumeJobPage.PageContent);
+            }
+            
+            Console.ReadKey();
+        }
+    }
+}
