+++ Observer Design Pattern/0011 - Behavioral - Observer Pattern.csproj
@@ -50,7 +50,7 @@
     <Compile Include="IProductObserver.cs" />
     <Compile Include="ProductNotifier.cs" />
     <Compile Include="ProductObserver.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="_Main.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
