+++ Singleton design pattern/0004 - Creational - Singleton pattern.csproj
@@ -48,7 +48,7 @@
   <ItemGroup>
     <Compile Include="ILogger.cs" />
     <Compile Include="Logger.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="_Main.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
